Refatoração geral
