+++ package.json
@@ -2,13 +2,13 @@
   "name": "back-end",
   "version": "1.0.0",
   "description": "",
-  "main": "index.js",
+  "main": "./src/index.js",
   "scripts": {
-    "pm2": "node ./utils/swagger-gen.js && pm2 start processes.json --no-daemon",
-    "start": "node ./utils/swagger-gen.js && node index.js",
-    "dev": " node ./utils/swagger-gen.js && nodemon index.js",
+    "pm2": "node ./src/utils/swagger-gen.js && pm2 start processes.json --no-daemon",
+    "start": "node ./src/utils/swagger-gen.js && node ./src/index.js",
+    "dev": " node ./src/utils/swagger-gen.js && nodemon ./src/index.js",
     "migrate": "sequelize-cli db:migrate",
-    "deploy": "node ./utils/swagger-gen.js && sequelize-cli db:migrate && pm2 start processes.json --no-daemon"
+    "deploy": "node ./src/utils/swagger-gen.js && sequelize-cli db:migrate && pm2 start processes.json --no-daemon"
   },
   "repository": {
     "type": "git",
